Figure5Clean
