+++ ckine/MBmodel.py
@@ -6,7 +6,7 @@ from os.path import dirname, join
 import numpy as np
 import pandas as pd
 from scipy.optimize import root
-from .imports import import_pstat_all, getBindDict, importReceptors
+from .imports import getBindDict, importReceptors
 
 path_here = dirname(dirname(__file__))
 
@@ -101,58 +101,38 @@ def cytBindingModel(mut, val, doseVec, cellType, x=False, date=False):
             output *= convDict.loc[(convDict.Date == date) & (convDict.Cell == cellType)].Scale.values
     return output
 
+def cytBindingModel_basicSelec(counts, x=False, date=False):
+    """Runs binding model for a given dataframe of epitope abundances"""
+    mut = 'IL2'
+    val = 1
+    doseVec = np.array([0.1])
+    #
+    recCount = np.ravel(counts)
+    #
+    mutAffDF = pd.read_csv(join(path_here, "ckine/data/WTmutAffData.csv"))
+    Affs = mutAffDF.loc[(mutAffDF.Mutein == mut)]
+    Affs = np.power(np.array([Affs["IL2RaKD"].values, Affs["IL2RBGKD"].values]) / 1e9, -1)
+    Affs = np.reshape(Affs, (1, -1))
+    Affs = np.repeat(Affs, 2, axis=0)
+    np.fill_diagonal(Affs, 1e2)  # Each cytokine can only bind one a and one b
 
-def runFullModel(x=False, time=[0.5], saveDict=False, singleCell=False):
-    """Runs model for all data points and outputs date conversion dict for binding to pSTAT. Can be used to fit Kx"""
-    statDF = import_pstat_all(singleCell)
-    statDF = statDF.loc[(statDF.Ligand != "H16L N-term (Mono)") & (statDF.Ligand != "IL15 (Mono)")]
-    statDF = statDF.loc[(statDF.Time.isin(time))]
+    if doseVec.size == 1:
+        doseVec = np.array([doseVec])
+    output = np.zeros(doseVec.size)
 
-    dateConvDF = pd.DataFrame(columns={"Date", "Scale", "Cell"})
-    masterSTAT = pd.DataFrame(columns={"Ligand", "Date", "Cell", "Time", "Dose", "Valency", "Experimental", "Predicted"})
-    dates = statDF.Date.unique()
-
-    for (date, lig, conc, cell, time), group in statDF.groupby(["Date", "Ligand", "Dose", "Cell", "Time"]):
-        if lig[-5::] == "(Biv)":
-            val = 2
-            ligName = lig[0:-6]
+    for i, dose in enumerate(doseVec):
+        if x:
+            output[i] = polyc(dose / 1e9, np.power(10, x[0]), recCount, [[val, val]], [1.0], Affs)[0][1]
         else:
-            val = 1
-            ligName = lig[0:-7]
+            output[i] = polyc(dose / 1e9, getKxStar(), recCount, [[val, val]], [1.0], Affs)[0][1]  # IL2RB binding only
 
-        entry = group.Mean.values
-        if len(entry) >= 1:
-            expVal = np.mean(entry)
-            # print(type(conc))
-            predVal = cytBindingModel(ligName, val, conc, cell, x)
-            masterSTAT = masterSTAT.append(pd.DataFrame({"Ligand": ligName, "Date": date, "Cell": cell, "Dose": conc,
-                                                         "Time": time, "Valency": val, "Experimental": expVal, "Predicted": predVal}))
-
-    for date in dates:
-        for cell in masterSTAT.Cell.unique():
-            if cell[-1] == "$":  # if it is a binned pop, use ave fit
-                predVecBin = masterSTAT.loc[(masterSTAT.Date == date) & (masterSTAT.Cell == cell)].Predicted.values
-                slope = dateConvDF.loc[(dateConvDF.Date == date) & (dateConvDF.Cell == cell[0:-13])].Scale.values
-                masterSTAT.loc[(masterSTAT.Date == date) & (masterSTAT.Cell == cell), "Predicted"] = predVecBin * slope
-            else:
-                expVec = masterSTAT.loc[(masterSTAT.Date == date) & (masterSTAT.Cell == cell)].Experimental.values
-                predVec = masterSTAT.loc[(masterSTAT.Date == date) & (masterSTAT.Cell == cell)].Predicted.values
-                slope = np.linalg.lstsq(np.reshape(predVec, (-1, 1)), np.reshape(expVec, (-1, 1)), rcond=None)[0][0]
-                masterSTAT.loc[(masterSTAT.Date == date) & (masterSTAT.Cell == cell), "Predicted"] = predVec * slope
-                dateConvDF = dateConvDF.append(pd.DataFrame({"Date": date, "Scale": slope, "Cell": cell}))
-    if saveDict:
-        dateConvDF.set_index("Date").to_csv(join(path_here, "ckine/data/BindingConvDict.csv"))
-
-    if x:
-        return np.linalg.norm(masterSTAT.Predicted.values - masterSTAT.Experimental.values)
-    else:
-        return masterSTAT
+    return output
 
 
 # CITEseq Tetra valent exploration functions below
 
 def cytBindingModel_CITEseq(counts, betaAffs, val, mut, x=False, date=False):
-    """Runs binding model for a given mutein, valency, dose, and cell type."""
+    """Runs binding model for a given epitopes abundance, betaAffinity, valency, and mutein type."""
 
     doseVec = np.array([0.1])
     recCount = np.ravel(counts)
@@ -185,7 +165,7 @@ def cytBindingModel_bispecCITEseq(counts, betaAffs, recXaff, val, mut, x=False):
     recXaff = np.power(10, recXaff)
     doseVec = np.array([0.1])
     recCount = np.ravel(counts)
-
+    
     mutAffDF = pd.read_csv(join(path_here, "ckine/data/WTmutAffData.csv"))
     Affs = mutAffDF.loc[(mutAffDF.Mutein == mut)]
     Affs = np.power(np.array([Affs["IL2RaKD"].values, [betaAffs]]) / 1e9, -1)
@@ -204,6 +184,40 @@ def cytBindingModel_bispecCITEseq(counts, betaAffs, recXaff, val, mut, x=False):
             output[i] = polyc(dose / (val * 1e9), np.power(10, x[0]), recCount, [[val, val, val]], [1.0], Affs)[0][1]
         else:
             output[i] = polyc(dose / (val * 1e9), getKxStar(), recCount, [[val, val, val]], [1.0], Affs)[0][1]
+
+    return output
+
+def cytBindingModel_bispecOpt(counts, recXaff, x=False):
+    """Runs binding model for a given mutein, valency, dose, and cell type."""
+
+    mut = 'IL2'
+    val = 1
+    doseVec = np.array([0.1])
+
+    recXaff = np.power(10, recXaff)
+
+    recCount = np.ravel(counts)
+
+    mutAffDF = pd.read_csv(join(path_here, "ckine/data/WTmutAffData.csv"))
+    Affs = mutAffDF.loc[(mutAffDF.Mutein == mut)]
+    Affs = np.power(np.array([Affs["IL2RaKD"].values, Affs["IL2RBGKD"].values]) / 1e9, -1)
+    Affs = np.reshape(Affs, (1, -1))
+    Affs = np.append(Affs, recXaff)
+    holder = np.full((3, 3), 1e2)
+    np.fill_diagonal(holder, Affs)
+    Affs = holder
+
+    # Check that values are in correct placement, can invert
+
+    if doseVec.size == 1:
+        doseVec = np.array([doseVec])
+    output = np.zeros(doseVec.size)
+
+    for i, dose in enumerate(doseVec):
+        if x:
+            output[i] = polyc(dose / (val * 1e9), np.power(10, x[0]), recCount, [[val, val, val]], [1.0], Affs)[0][1]
+        else:
+            output[i] = polyc(dose / (val * 1e9), getKxStar(), recCount, [[val, val, val]], [1.0], Affs)[0][1]  # IL2RB binding only
 
     return output
