+++ ckine/selectivityFuncs.py
@@ -3,11 +3,10 @@ Functions used in binding and selectivity analysis
 """
 import string
 from .imports import importCITE, importReceptors
-from .MBmodel import cytBindingModel_CITEseq, cytBindingModel_bispecCITEseq
+from .MBmodel import cytBindingModel_CITEseq, cytBindingModel_bispecCITEseq, cytBindingModel_bispecOpt, cytBindingModel_basicSelec
 from os.path import dirname, join
 from scipy.optimize import minimize, Bounds
 import pandas as pd
-import seaborn as sns
 import numpy as np
 
 path_here = dirname(dirname(__file__))
@@ -32,7 +31,7 @@ def getSampleAbundances(epitopes: list, cellList: list):
 
     # Get conv factors, average them to use on epitopes with unlisted conv facts
     convFact = convFactCalc()
-    # Change to create labeled convFact df to use below (more transparent)
+    # TODO Change to create labeled convFact df to use below (more transparent) 
     meanConv = convFact.Weight.mean()
 
     # Sample sizes generated corresponding to cell list using mean values
@@ -61,6 +60,8 @@ def getSampleAbundances(epitopes: list, cellList: list):
                 convFact = 77.136987  # The values are from convFactCalc
             elif e == 'CD122':
                 convFact = 332.680090
+            elif e == "CD127":
+                convFact = 594.379215
             else:
                 convFact = meanConv
 
@@ -75,7 +76,7 @@ def getSampleAbundances(epitopes: list, cellList: list):
     return epitopesDF
 
 
-def getSignaling(betaAffs: np.array, targCell: string, offTCells: list, epitopesDF: pd.dataframe):
+def getSignaling(betaAffs: np.array, targCell: string, offTCells: list, epitopesDF: pd.DataFrame):
     """Returns total signaling summed over single cells for given parameters, can be adjusted for various purposes
     Args:
         betaAffs: 1D numpy array of beta receptor affinities over which to analyze signaling
@@ -118,7 +119,7 @@ def getSignaling(betaAffs: np.array, targCell: string, offTCells: list, epitopes
     return target_sigs, offTarg_sigs
 
 
-def bindingCalc(df: pd.dataframe, targCell: string, offTCells: list, betaAffs: np.array, val: int, mut: string, bispec=False, epitope: string=None):
+def bindingCalc(df: pd.DataFrame, targCell: string, offTCells: list, betaAffs: np.array, val: int, mut: string, bispec=False, epitope: string=None):
     """Calculates selectivity for no additional epitope
     Args:
         df: (epitopesDF) contains info on epitope abundance by cell type (generated by getSampleAbundance())
@@ -171,7 +172,7 @@ def bindingCalc(df: pd.dataframe, targCell: string, offTCells: list, betaAffs: n
     return targetBound, offTargetBound
 
 
-def minSelecFunc(x: float, df: pd.dataframe, targCell: string, offTCells: list, epitope: string):
+def minSelecFunc(x: float, df: pd.DataFrame, targCell: string, offTCells: list, epitope: string):
     """Serves as the function which will have its return value minimized to get optimal selectivity
     To be used in conjunction with optimizeDesign()
     Args:
@@ -197,21 +198,21 @@ def minSelecFunc(x: float, df: pd.dataframe, targCell: string, offTCells: list, 
         cd25Count = cd25DF[targCell].item()[i]
         cd122Count = cd122DF[targCell].item()[i]
         counts = [cd25Count, cd122Count, epCount]
-        targetBound += cytBindingModel_bispecCITEseq(counts, recXaff)
+        targetBound += cytBindingModel_bispecOpt(counts, recXaff)
     for cellT in offTCells:
         for i, epCount in enumerate(epitopeDF[cellT].item()):
             cd25Count = cd25DF[cellT].item()[i]
             cd122Count = cd122DF[cellT].item()[i]
             counts = [cd25Count, cd122Count, epCount]
 
-            offTargetBound += cytBindingModel_bispecCITEseq(counts, recXaff)
+            offTargetBound += cytBindingModel_bispecOpt(counts, recXaff)
 
     selectivity = (offTargetBound) / (targetBound)
 
     return selectivity
 
 
-def optimizeDesign(targCell: string, offTcells: list, selectedDF: pd.dataframe, epitope: string):
+def optimizeDesign(targCell: string, offTcells: list, selectedDF: pd.DataFrame, epitope: string):
     """ A general purpose optimizer used to minimize selectivity output by varying affinity parameter.
     Args:
         targCell: string cell type which is target and signaling is desired (basis of selectivity)
@@ -230,9 +231,36 @@ def optimizeDesign(targCell: string, offTcells: list, selectedDF: pd.dataframe, 
     optBnds = Bounds(np.full_like(X0, 6.0), np.full_like(X0, 9.0))
 
     optimized = minimize(minSelecFunc, X0, bounds=optBnds, args=(selectedDF, targCell, offTcells, epitope), jac="3-point")
-    optSelectivity = optimized.fun[0]
+    optSelectivity = optimized.fun
 
     return optSelectivity
+
+def selecCalc(df: pd.DataFrame, targCell: string, offTCells: list):
+    """Calculates selectivity for no additional epitope
+    Args:
+        targCell: string cell type which is target and signaling is desired (basis of selectivity)
+        offTCells: list of strings of cell types for which signaling is undesired
+        df: contains epitope abundance information by cell type(importantly including cd25 anf cd122 here)
+    Returns:
+        selectivity value of standard IL2 based on abundances
+    """
+    targetBound = 0
+    offTargetBound = 0
+
+    cd25DF = df.loc[(df.Epitope == 'CD25')]
+    cd122DF = df.loc[(df.Epitope == 'CD122')]
+
+    for i, cd25Count in enumerate(cd25DF[targCell].item()):
+        cd122Count = cd122DF[targCell].item()[i]
+        counts = [cd25Count, cd122Count]
+        targetBound += cytBindingModel_basicSelec(counts)
+    for cellT in offTCells:
+        for i, cd25Count in enumerate(cd25DF[cellT].item()):
+            cd122Count = cd122DF[cellT].item()[i]
+            counts = [cd25Count, cd122Count]
+            offTargetBound += cytBindingModel_basicSelec(counts)
+
+    return (offTargetBound) / (targetBound)
 
 
 cellDict = {"CD4 Naive": "Thelper",
